+++ client/package.json
@@ -14,9 +14,9 @@
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
+    "firebase": "^9.14.0",
     "prop-types": "^15.8.1",
     "qrcode": "^1.5.1",
-    "firebase": "^9.18.0",
     "jwt-decode": "^3.1.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
