+++ client/src/components/callteacher/Callteacher.scss
@@ -23,8 +23,6 @@
 .DivQr {
   display: flex;
   align-items: center;
-  margin: 2rem;
-  margin-left: 3rem;
 }
 
 .Qrcode {
