+++ server/package.json
@@ -19,6 +19,7 @@
     "express": "^4.18.2",
     "firebase": "^9.18.0",
     "firebase-admin": "^11.3.0",
+    "http": "^0.0.1-security",
     "jsonwebtoken": "^9.0.0",
     "react-dotenv": "^0.1.3",
     "ws": "^8.11.0"
